Split peer health refresh tests
