+++ .pm/tasks/task_a7b607e4bf1a4392b8ad62297b726d55.execution.md
@@ -140,3 +140,25 @@ Example:
 - Expected Result: task verification passes and closeout records done status; repository-wide PM lint may fail only on unrelated historical debt.
 - Actual Result: partial pass; `./scripts/cargo-dev.sh test -p oasis7_net` passed and task YAML recorded done/verified/closed, while final repo-wide `pm-lint` failed on existing unrelated historical execution-log formatting debt such as task_04d61dc5778e4b1683a61056daf454e3 and task_060e9de147ba4757ac29cf0fb7a15210.
 - Blocker / Next Action: run `./scripts/pm/workflow-lint.sh --phase pr-ready --task-uid task_a7b607e4bf1a4392b8ad62297b726d55`; if task-local pr-ready lint passes, continue PR creation.
+
+## 2026-06-27 20:20:36 CST / tpm
+- 完成内容: Investigated PR #708 required-gate failure. GitHub Actions failed in `./scripts/check-rust-file-size.sh` because `crates/oasis7_net/src/libp2p_net/tests.rs` reached 1233 lines, exceeding the 1200-line Rust file-size gate after the constrained peer-health regression was added.
+- 完成内容: Fixed the gate issue without changing production behavior by moving the three `refresh_peer_manager_healths*` regressions into `crates/oasis7_net/src/libp2p_net/tests/peer_health_refresh_tests.rs` and adding the test module declaration from `crates/oasis7_net/src/libp2p_net/tests.rs`.
+- 遗留事项: commit and push the file-size remediation, then continue PR required-check watch.
+- Action: rerun local verification after splitting peer-health refresh tests out of the oversized test file.
+- Validation Command: `env -u RUSTC_WRAPPER cargo fmt --all -- --check`
+- Expected Result: formatting stays canonical after moving the tests.
+- Actual Result: passed.
+- Validation Command: `./scripts/check-rust-file-size.sh`
+- Expected Result: no oversized Rust source/test files remain for the current branch.
+- Actual Result: passed; oversized code/test/structural scan counts were 0.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7_net refresh_peer_manager_healths`
+- Expected Result: moved focused peer-health refresh tests still pass.
+- Actual Result: passed; 3 passed, 0 failed.
+- Validation Command: `./scripts/cargo-dev.sh test -p oasis7_net`
+- Expected Result: full `oasis7_net` crate verification still passes after the test-module split.
+- Actual Result: passed; 171 passed, 0 failed, doc-tests 0 passed.
+- Validation Command: `git diff --check`; `./scripts/pm/workflow-lint.sh --phase current --task-uid task_a7b607e4bf1a4392b8ad62297b726d55`; `./scripts/doc-governance-check.sh`; `./scripts/pm/required-tier-smoke.sh`
+- Expected Result: diff hygiene, task-local workflow lint, doc governance, and required-tier smoke remain green.
+- Actual Result: passed; all listed gates passed.
+- Blocker / Next Action: commit, push, and watch PR #708 checks.
